move logic to resolve event ids based on state to the EventId class
