+++ src/commonMain/kotlin/org/cru/godtools/tool/state/State.kt
@@ -2,8 +2,6 @@ package org.cru.godtools.tool.state
 
 import org.cru.godtools.tool.internal.Parcelable
 import org.cru.godtools.tool.internal.Parcelize
-import org.cru.godtools.tool.model.EVENT_NAMESPACE_STATE
-import org.cru.godtools.tool.model.EventId
 
 @Parcelize
 class State internal constructor(private val state: MutableMap<String, List<String>>) : Parcelable {
@@ -15,9 +13,4 @@ class State internal constructor(private val state: MutableMap<String, List<Stri
     internal operator fun set(key: String, value: String?) = state.set(key, listOfNotNull(value))
     internal operator fun set(key: String, values: List<String>?) = state.set(key, values.orEmpty())
     internal fun set(key: String, vararg values: String) = state.set(key, values.toList())
-
-    fun resolveEventId(id: EventId) = when (id.namespace) {
-        EVENT_NAMESPACE_STATE -> getAll(id.name).map { EventId(name = it) }
-        else -> listOf(id)
-    }
 }
